order_queue/requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.2

watchdog==6.0.0
openai==1.60.0  # Ensure OpenAI API library is installed

requests==2.26.0

